chore(crc): use boost CRC, move to util
