+++ nebula_decoders/src/nebula_decoders_continental/decoders/continental_srr520_decoder.cpp
@@ -15,7 +15,7 @@
 #include "nebula_decoders/nebula_decoders_continental/decoders/continental_srr520_decoder.hpp"
 
 #include <nebula_common/continental/continental_srr520.hpp>
-#include <nebula_common/continental/crc.hpp>
+#include <nebula_common/util/crc.hpp>
 
 #include <boost/algorithm/string/join.hpp>
